+++ core-processor/src/main/java/jeff/core/aop/ControllerExceptionResAspect.java
@@ -3,7 +3,7 @@ package jeff.core.aop;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jeff.common.consts.ResponseCode;
-import jeff.common.entity.dto.send.ResponseObject;
+import jeff.common.entity.dto.receive.ResponseObjectFromInnerSystem;
 import jeff.common.util.LogUtil;
 import jeff.common.exception.BusyException;
 import jeff.common.exception.MyNotFoundException;
@@ -11,6 +11,7 @@ import jeff.core.exception.OrderException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
@@ -19,7 +20,9 @@ import javax.annotation.PostConstruct;
 
 /**
  * 切面程式，攔截控制器所拋出的例外，根據例外去回應給客戶端不同的訊息。
- * 遭遇錯誤，統一也回Http Code200，會有自訂義的code去表示請求的狀態。
+ *
+ * <p><del>遭遇錯誤，統一也回Http Code200，會有自訂義的code去表示請求的狀態。</del></p> <---如果這樣做，那麼上游Server的resilience4j不會把錯誤的請求算到斷路器count裡面。
+ * 遭遇錯誤時，回應Http狀態碼為非2XX。
  */
 @Slf4j
 @RestControllerAdvice(basePackages = {"jeff.core.controller"})
@@ -46,29 +49,32 @@ public class ControllerExceptionResAspect {
 
 
     @ExceptionHandler(value = OrderException.class)
-    @ResponseStatus(code = HttpStatus.OK)
-    public ResponseObject handleOrderException(OrderException oe) {
-        return new ResponseObject(ResponseCode.Failure.getCode(), EMPTY_CONTENT, oe.getMessage());
+    public ResponseEntity<ResponseObjectFromInnerSystem> handleOrderException(OrderException oe) {
+        return ResponseEntity
+                .status(ResponseCode.Failure.getCode())
+                .body(new ResponseObjectFromInnerSystem(ResponseCode.Failure.getCode(), EMPTY_CONTENT, oe.getMessage()));
     }
 
     @ExceptionHandler(value = MyNotFoundException.class)
-    @ResponseStatus(code = HttpStatus.OK)
-    public ResponseObject handleMyNotFoundException(MyNotFoundException mnfe) {
-        return new ResponseObject(ResponseCode.NotFound.getCode(), EMPTY_CONTENT, mnfe.getMessage());
+    public ResponseEntity<ResponseObjectFromInnerSystem> handleMyNotFoundException(MyNotFoundException mnfe) {
+        return ResponseEntity
+                .status(ResponseCode.NotFound.getCode())
+                .body(new ResponseObjectFromInnerSystem(ResponseCode.NotFound.getCode(), EMPTY_CONTENT, mnfe.getMessage()));
     }
 
     @ExceptionHandler(value = BusyException.class)
-    @ResponseStatus(code = HttpStatus.OK)
-    public ResponseObject handleBusyException(BusyException be) {
-        return new ResponseObject(ResponseCode.TooManyReq.getCode(), EMPTY_CONTENT, be.getMessage());
+    public ResponseEntity<ResponseObjectFromInnerSystem> handleBusyException(BusyException be) {
+        return ResponseEntity
+                .status(ResponseCode.TooManyReq.getCode())
+                .body(new ResponseObjectFromInnerSystem(ResponseCode.TooManyReq.getCode(), EMPTY_CONTENT, be.getMessage()));
     }
 
     /**
      * 當發生了一些沒有預期到的例外。
      */
     @ExceptionHandler(value = Exception.class)
-    @ResponseStatus(code = HttpStatus.OK)
-    public ResponseObject handleException(Exception e) {
+    @ResponseStatus(code = HttpStatus.INTERNAL_SERVER_ERROR)
+    public ResponseObjectFromInnerSystem handleException(Exception e) {
         logUtil.logError(
                 log,
                 logUtil.composeLogPrefixForSystem(),
@@ -76,7 +82,7 @@ public class ControllerExceptionResAspect {
                 e
         );
 
-        return new ResponseObject(ResponseCode.Failure.getCode(), EMPTY_CONTENT, "Some errors occurred while processing request, please call the application owner.");
+        return new ResponseObjectFromInnerSystem(ResponseCode.Failure.getCode(), EMPTY_CONTENT, "Some errors occurred while processing request, please call the application owner.");
     }
 
 }