+++ high-concurrency-processor/src/main/java/jeff/highconcurrency/aop/ControllerExceptionResAspect.java
@@ -3,13 +3,15 @@ package jeff.highconcurrency.aop;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jeff.common.consts.ResponseCode;
-import jeff.common.entity.dto.send.ResponseObject;
+import jeff.common.entity.dto.receive.ResponseObjectFromInnerSystem;
+import jeff.common.exception.MyInnerCommunicationStatusFailureException;
 import jeff.common.util.LogUtil;
 import jeff.common.exception.BusyException;
 import jeff.highconcurrency.exception.FlashSaleEventConsumeException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
@@ -18,7 +20,9 @@ import javax.annotation.PostConstruct;
 
 /**
  * 切面程式，攔截控制器所拋出的例外，根據例外去回應給客戶端不同的訊息。
- * 遭遇錯誤，統一也回Http Code200，會有自訂義的code去表示請求的狀態。
+ *
+ *  <p><del>遭遇錯誤，統一也回Http Code200，會有自訂義的code去表示請求的狀態。</del></p> <---如果這樣做，那麼上游Server的resilience4j不會把錯誤的請求算到斷路器count裡面。
+ *  遭遇錯誤時，回應Http狀態碼為非2XX。
  */
 @Slf4j
 @RestControllerAdvice(basePackages = {"jeff.highconcurrency.controller"})
@@ -44,23 +48,34 @@ public class ControllerExceptionResAspect {
     }
 
     @ExceptionHandler(value = FlashSaleEventConsumeException.class)
-    @ResponseStatus(code = HttpStatus.OK)
-    public ResponseObject handleOrderException(FlashSaleEventConsumeException fsece) {
-        return new ResponseObject(ResponseCode.Failure.getCode(), EMPTY_CONTENT, fsece.getMessage());
+    public ResponseEntity<ResponseObjectFromInnerSystem> handleOrderException(FlashSaleEventConsumeException fsece) {
+        return ResponseEntity
+                .status(ResponseCode.Failure.getCode())
+                .body(new ResponseObjectFromInnerSystem(ResponseCode.Failure.getCode(), EMPTY_CONTENT, fsece.getMessage()));
     }
 
     @ExceptionHandler(value = BusyException.class)
-    @ResponseStatus(code = HttpStatus.OK)
-    public ResponseObject handleBusyException(BusyException be) {
-        return new ResponseObject(ResponseCode.TooManyReq.getCode(), EMPTY_CONTENT, be.getMessage());
+    public ResponseEntity<ResponseObjectFromInnerSystem> handleBusyException(BusyException be) {
+        return ResponseEntity
+                .status(ResponseCode.TooManyReq.getCode())
+                .body(new ResponseObjectFromInnerSystem(ResponseCode.TooManyReq.getCode(), EMPTY_CONTENT, be.getMessage()));
     }
+
+    @ExceptionHandler(value = MyInnerCommunicationStatusFailureException.class)
+    public ResponseEntity<ResponseObjectFromInnerSystem> handleMyInnerCommunicationStatusFailureException(MyInnerCommunicationStatusFailureException micsfe) {
+        return ResponseEntity
+                .status(micsfe.getStatusCode())
+                .body(micsfe.getResObjectFromInnerSystem());
+    }
+
+
 
     /**
      * 當發生了一些沒有預期到的例外。
      */
     @ExceptionHandler(value = Exception.class)
-    @ResponseStatus(code = HttpStatus.OK)
-    public ResponseObject handleException(Exception e) {
+    @ResponseStatus(code = HttpStatus.INTERNAL_SERVER_ERROR)
+    public ResponseObjectFromInnerSystem handleException(Exception e) {
         logUtil.logError(
                 log,
                 logUtil.composeLogPrefixForSystem(),
@@ -68,7 +83,7 @@ public class ControllerExceptionResAspect {
                 e
         );
 
-        return new ResponseObject(ResponseCode.Failure.getCode(), EMPTY_CONTENT, "Some errors occurred while processing request, please call the application owner.");
+        return new ResponseObjectFromInnerSystem(ResponseCode.Failure.getCode(), EMPTY_CONTENT, "Some errors occurred while processing request, please call the application owner.");
     }
 
 }